// src/modules/clientes/types/Cliente.ts
export interface Cliente {
  idCliente?: string
  descCliente: string
  razonSocial: string
  cifCliente: string
  direccion: string
  codPostal: string
  poblacion: string
  provincia: string
  pais: string
  telefono: string
  fax: string
  eMail: string
  web: string
  telefono2: string
  movil: string
  formaPago: string
  ccCliente: string
  dtoComercial: number
  observaciones: string
}